backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
silero-vad
ffmpeg-python
colorlog
numpy
websockets
scipy
soundfile
